requirements.txt
lxml
wget
gtts
beautifulsoup4
requests
python-telegram-bot==13.9.0
SQLAlchemy==1.4.28
psycopg2-binary
feedparser
faker
hachoir
pykeyboard
pynewtonmath
aiofiles
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow>=9.0
CurrencyConverter
googletrans
google-trans-new
speedtest-cli
coffeehouse
regex
bleach
wikipedia
markdown2
telethon
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub
TgCrypto
Pyrogram
PyDictionary
git+https://github.com/Sorrybabe/downloader
youtube_search_python
lyrics_extractor
asyncio
cloudscraper
dateparser
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
redis
janda
pySmartDL
bing_image_downloader
search_engine_parser
pytz
jikanpy
tswift
python-arq
envparse
#profanity (Credits to Julia)
better_profanity
textblob 
nudepy
nekos.py
motor
gpytranslate
fontTools
opencv-python-headless
odmantic
tzlocal
pendulum
fuzzysearch
img2pdf
multicolorcaptcha
git+https://github.com/python-telegram-bot/ptbcontrib.git
PyYAML>5.0
loguru
aiofiles
pytube
yt_dlp
pytimeparse
youtube_dl
